Replaced hardcoded list with data from DB via repository
The hardcoded list returned by the '/books' endpoint of the rest
controller is replaced with data from the database. For this, a
repository layer is added (BookRepository). A service layer is not
yet necessary since no business logic is being added.

Next steps are to create more endpoints for database queries, to
allow POSTING data and also more granular queries.