Updated error handing for book insertion
To relay more descriptive error messages to the client, specifically
in the case of invalid inputs (e.g. blank strings or duplicates)
the exception handing of the book repository `addBook` method is
updated and the controller and tests are changed accordingly.
A unique constraint is added on the book name and a corresponding
test.